Move jujud related errors from utils package to cmd/jujud/agent/errors package;
