+++ cmd/jujud/agent/checkconnection.go
@@ -14,7 +14,7 @@ import (
 	"github.com/juju/juju/agent"
 	"github.com/juju/juju/api"
 	jujucmd "github.com/juju/juju/cmd"
-	"github.com/juju/juju/cmd/jujud/util"
+	agenterrors "github.com/juju/juju/cmd/jujud/agent/errors"
 	"github.com/juju/juju/worker/apicaller"
 )
 
@@ -55,7 +55,7 @@ func (c *checkConnectionCommand) Info() *cmd.Info {
 // Init is part of cmd.Command.
 func (c *checkConnectionCommand) Init(args []string) error {
 	if len(args) == 0 {
-		return &util.FatalError{"agent-name argument is required"}
+		return &agenterrors.FatalError{"agent-name argument is required"}
 	}
 	agentName, args := args[0], args[1:]
 	if err := cmd.CheckEmpty(args); err != nil {
@@ -66,7 +66,7 @@ func (c *checkConnectionCommand) Init(args []string) error {
 		return errors.Annotatef(err, "agent-name")
 	}
 	if tag.Kind() != "machine" && tag.Kind() != "unit" {
-		return &util.FatalError{"agent-name must be a machine or unit tag"}
+		return &agenterrors.FatalError{"agent-name must be a machine or unit tag"}
 	}
 	err = c.config.ReadConfig(agentName)
 	if err != nil {
