+++ cmd/jujud/agent/errors/errors_test.go
@@ -1,19 +1,20 @@
-// Copyright 2014 Canonical Ltd.
+// Copyright 2020 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package util
+package errors_test
 
 import (
 	stderrors "errors"
 
 	"github.com/juju/errors"
+	"github.com/juju/loggo"
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/apiserver/params"
+	agenterrors "github.com/juju/juju/cmd/jujud/agent/errors"
 	coretesting "github.com/juju/juju/testing"
 	"github.com/juju/juju/worker"
-	"github.com/juju/juju/worker/upgrader"
 )
 
 var (
@@ -22,6 +23,12 @@ var (
 
 type toolSuite struct {
 	coretesting.BaseSuite
+	logger loggo.Logger
+}
+
+func (s *toolSuite) SetUpTest(c *gc.C) {
+	s.BaseSuite.SetUpTest(c)
+	s.logger = loggo.GetLogger("test")
 }
 
 func (*toolSuite) TestErrorImportance(c *gc.C) {
@@ -29,16 +36,16 @@ func (*toolSuite) TestErrorImportance(c *gc.C) {
 	errorImportanceTests := []error{
 		nil,
 		stderrors.New("foo"),
-		&upgrader.UpgradeReadyError{},
+		&agenterrors.UpgradeReadyError{},
 		worker.ErrTerminateAgent,
 	}
 
 	for i, err0 := range errorImportanceTests {
 		for j, err1 := range errorImportanceTests {
-			c.Assert(MoreImportant(err0, err1), gc.Equals, i > j)
+			c.Assert(agenterrors.MoreImportant(err0, err1), gc.Equals, i > j)
 
 			// Should also work if errors are wrapped.
-			c.Assert(MoreImportant(errors.Trace(err0), errors.Trace(err1)), gc.Equals, i > j)
+			c.Assert(agenterrors.MoreImportant(errors.Trace(err0), errors.Trace(err1)), gc.Equals, i > j)
 		}
 	}
 }
@@ -60,7 +67,7 @@ var isFatalTests = []struct {
 		err:     errors.Trace(worker.ErrRestartAgent),
 		isFatal: true,
 	}, {
-		err:     &upgrader.UpgradeReadyError{},
+		err:     &agenterrors.UpgradeReadyError{},
 		isFatal: true,
 	}, {
 		err: &params.Error{
@@ -69,7 +76,7 @@ var isFatalTests = []struct {
 		},
 		isFatal: false,
 	}, {
-		err:     &FatalError{"some fatal error"},
+		err:     &agenterrors.FatalError{"some fatal error"},
 		isFatal: true,
 	}, {
 		err:     stderrors.New("foo"),
@@ -90,7 +97,7 @@ func (s *toolSuite) TestConnectionIsFatal(c *gc.C) {
 	for i, conn := range []*testConn{errConn, okConn} {
 		for j, test := range isFatalTests {
 			c.Logf("test %d.%d: %s", i, j, test.err)
-			fatal := ConnectionIsFatal(logger, conn)(test.err)
+			fatal := agenterrors.ConnectionIsFatal(s.logger, conn)(test.err)
 			if test.isFatal {
 				c.Check(fatal, jc.IsTrue)
 			} else {
@@ -106,15 +113,15 @@ func (s *toolSuite) TestConnectionIsFatalWithMultipleConns(c *gc.C) {
 
 	someErr := stderrors.New("foo")
 
-	c.Assert(ConnectionIsFatal(logger, okConn, okConn)(someErr),
+	c.Assert(agenterrors.ConnectionIsFatal(s.logger, okConn, okConn)(someErr),
 		jc.IsFalse)
-	c.Assert(ConnectionIsFatal(logger, okConn, okConn, okConn)(someErr),
+	c.Assert(agenterrors.ConnectionIsFatal(s.logger, okConn, okConn, okConn)(someErr),
 		jc.IsFalse)
-	c.Assert(ConnectionIsFatal(logger, okConn, errConn)(someErr),
+	c.Assert(agenterrors.ConnectionIsFatal(s.logger, okConn, errConn)(someErr),
 		jc.IsTrue)
-	c.Assert(ConnectionIsFatal(logger, okConn, okConn, errConn)(someErr),
+	c.Assert(agenterrors.ConnectionIsFatal(s.logger, okConn, okConn, errConn)(someErr),
 		jc.IsTrue)
-	c.Assert(ConnectionIsFatal(logger, errConn, okConn, okConn)(someErr),
+	c.Assert(agenterrors.ConnectionIsFatal(s.logger, errConn, okConn, okConn)(someErr),
 		jc.IsTrue)
 }
 
@@ -128,7 +135,7 @@ func (s *toolSuite) TestPingerIsFatal(c *gc.C) {
 	for i, pinger := range []testPinger{errPinger, okPinger} {
 		for j, test := range isFatalTests {
 			c.Logf("test %d.%d: %s", i, j, test.err)
-			fatal := PingerIsFatal(logger, pinger)(test.err)
+			fatal := agenterrors.PingerIsFatal(s.logger, pinger)(test.err)
 			if test.isFatal {
 				c.Check(fatal, jc.IsTrue)
 			} else {
@@ -145,18 +152,16 @@ func (s *toolSuite) TestPingerIsFatalWithMultipleConns(c *gc.C) {
 	var okPinger testPinger = func() error {
 		return nil
 	}
-
 	someErr := stderrors.New("foo")
-
-	c.Assert(PingerIsFatal(logger, okPinger, okPinger)(someErr),
+	c.Assert(agenterrors.PingerIsFatal(s.logger, okPinger, okPinger)(someErr),
 		jc.IsFalse)
-	c.Assert(PingerIsFatal(logger, okPinger, okPinger, okPinger)(someErr),
+	c.Assert(agenterrors.PingerIsFatal(s.logger, okPinger, okPinger, okPinger)(someErr),
 		jc.IsFalse)
-	c.Assert(PingerIsFatal(logger, okPinger, errPinger)(someErr),
+	c.Assert(agenterrors.PingerIsFatal(s.logger, okPinger, errPinger)(someErr),
 		jc.IsTrue)
-	c.Assert(PingerIsFatal(logger, okPinger, okPinger, errPinger)(someErr),
+	c.Assert(agenterrors.PingerIsFatal(s.logger, okPinger, okPinger, errPinger)(someErr),
 		jc.IsTrue)
-	c.Assert(PingerIsFatal(logger, errPinger, okPinger, okPinger)(someErr),
+	c.Assert(agenterrors.PingerIsFatal(s.logger, errPinger, okPinger, okPinger)(someErr),
 		jc.IsTrue)
 }
 
@@ -164,7 +169,7 @@ func (*toolSuite) TestIsFatal(c *gc.C) {
 
 	for i, test := range isFatalTests {
 		c.Logf("test %d: %s", i, test.err)
-		c.Assert(IsFatal(test.err), gc.Equals, test.isFatal)
+		c.Assert(agenterrors.IsFatal(test.err), gc.Equals, test.isFatal)
 	}
 }
