+++ cmd/jujud/agent/model.go
@@ -20,6 +20,8 @@ import (
 	"github.com/juju/juju/caas"
 	caasprovider "github.com/juju/juju/caas/kubernetes/provider"
 	jujucmd "github.com/juju/juju/cmd"
+	"github.com/juju/juju/cmd/jujud/agent/engine"
+	agenterrors "github.com/juju/juju/cmd/jujud/agent/errors"
 	"github.com/juju/juju/cmd/jujud/agent/modeloperator"
 	cmdutil "github.com/juju/juju/cmd/jujud/util"
 	jujuversion "github.com/juju/juju/version"
@@ -54,7 +56,7 @@ func (m *ModelCommand) Info() *cmd.Info {
 // Init initializers the command for running
 func (m *ModelCommand) Init(args []string) error {
 	if m.ModelUUID == "" {
-		return cmdutil.RequiredError("model-uuid")
+		return agenterrors.RequiredError("model-uuid")
 	}
 
 	if err := m.AgentConf.CheckArgs(args); err != nil {
@@ -62,8 +64,8 @@ func (m *ModelCommand) Init(args []string) error {
 	}
 
 	m.runner = worker.NewRunner(worker.RunnerParams{
-		IsFatal:       cmdutil.IsFatal,
-		MoreImportant: cmdutil.MoreImportant,
+		IsFatal:       agenterrors.IsFatal,
+		MoreImportant: agenterrors.MoreImportant,
 		RestartDelay:  jworker.RestartDelay,
 	})
 	return nil
@@ -139,7 +141,7 @@ func (m *ModelCommand) Workers() (worker.Worker, error) {
 		NewContainerBrokerFunc: caas.New,
 	})
 
-	engine, err := dependency.NewEngine(DependencyEngineConfig())
+	engine, err := dependency.NewEngine(engine.DependencyEngineConfig())
 	if err != nil {
 		return nil, err
 	}
