+++ cmd/jujud/agent/unit.go
@@ -22,6 +22,8 @@ import (
 	"github.com/juju/juju/api/base"
 	"github.com/juju/juju/api/uniter"
 	jujucmd "github.com/juju/juju/cmd"
+	"github.com/juju/juju/cmd/jujud/agent/engine"
+	agenterrors "github.com/juju/juju/cmd/jujud/agent/errors"
 	"github.com/juju/juju/cmd/jujud/agent/unit"
 	cmdutil "github.com/juju/juju/cmd/jujud/util"
 	"github.com/juju/juju/core/machinelock"
@@ -97,7 +99,7 @@ func (a *UnitAgent) SetFlags(f *gnuflag.FlagSet) {
 // Init initializes the command for running.
 func (a *UnitAgent) Init(args []string) error {
 	if a.UnitName == "" {
-		return cmdutil.RequiredError("unit-name")
+		return agenterrors.RequiredError("unit-name")
 	}
 	if !names.IsValidUnit(a.UnitName) {
 		return errors.Errorf(`--unit-name option expects "<application>/<n>" argument`)
@@ -106,8 +108,8 @@ func (a *UnitAgent) Init(args []string) error {
 		return err
 	}
 	a.runner = worker.NewRunner(worker.RunnerParams{
-		IsFatal:       cmdutil.IsFatal,
-		MoreImportant: cmdutil.MoreImportant,
+		IsFatal:       agenterrors.IsFatal,
+		MoreImportant: agenterrors.MoreImportant,
 		RestartDelay:  jworker.RestartDelay,
 	})
 
@@ -207,7 +209,7 @@ func (a *UnitAgent) APIWorkers() (worker.Worker, error) {
 		Clock:                clock.WallClock,
 	})
 
-	engine, err := dependency.NewEngine(DependencyEngineConfig())
+	engine, err := dependency.NewEngine(engine.DependencyEngineConfig())
 	if err != nil {
 		return nil, err
 	}
