+++ cmd/k8sagent/unit/manifolds.go
@@ -1,4 +1,4 @@
-// Copyright 2012-2020 Canonical Ltd.
+// Copyright 2020 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
 package unit
